README
There are two folders, one for MiniMax and one for MonteCarlo. These folders include 3 files, 1 main, 1 mancala-kalaha and one for the respective AI algorithm used
In order to run the program what we need to do is open the folder with Visual Studio Code or Pycharm or any other software used for python and run the main, the imports needed are already there.
The programming language used is Python.